backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==3.0.1
openpyxl==3.1.5
sqlalchemy==2.0.36
python-dotenv==1.0.1
pydantic==1.10.17
